+++ ServerData.cs
@@ -141,5 +141,5 @@ public record RouteInfo(string Route, string Method, bool IsLegacy = false);
 
 public static class APIVersion
 {
-    public const string Version = "1.6.3";
+    public const string Version = "1.6.4";
 }